+++ src/editors/array-table-object.js
@@ -140,6 +140,7 @@ class EditorArrayTableObject extends EditorArray {
     })
 
     this.refreshSortable(table.tbody)
+    this.refreshAddBtn()
     this.refreshDisabledState()
     this.refreshScrollPosition(table.container)